% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVar.R
\name{scVar}
\alias{scVar}
\title{Identifying biological variability in scRNA-seq data}
\usage{
scVar(
  dat,
  norm,
  metric = c("SD", "MAD", "IQR", "CV", "FF", "scran", "LCV", "DESeq2", "edgeR",
    "Seurat_mvp", "Seurat_vst", "DM", "glmGamPoi", "BASiCS"),
  org = "mm9",
  group = NULL
)
}
\arguments{
\item{dat}{Raw count matrix}

\item{norm}{Logic parameter to indicate normalisation (default run by scran)}

\item{metric}{Name of the metric to run for analysis}

\item{org}{Organisms name, only required for DM}

\item{group}{covariates for the count matrix.}
}
\value{
A named vector with corresponding gene expression variability
}
\description{
Identifying biological variability in scRNA-seq data
}
\author{
Huiwen Zheng
}
